pyproject.toml
[project]
name = "event-bridge"
version = "0.1.0"
description = "Self-hosted inbound webhook gateway with forwarding, retry and dead-letter queue."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "restrictedpython>=8.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "anyio>=4.0.0",
]
